+++ src/commons/sideContent/content/SideContentCanvasOutput.tsx
