+++ src/commons/sideContent/content/SideContentContestVotingContainer.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { showWarningMessage } from 'src/commons/utils/notifications/NotificationsHelper';
 
-import { ContestEntry } from '../assessment/AssessmentTypes';
+import { ContestEntry } from '../../assessment/AssessmentTypes';
 import SideContentContestVoting from './SideContentContestVoting';
 
 export type SideContentContestVotingContainerProps = DispatchProps & StateProps;
