+++ src/commons/sideContent/content/SideContentDataVisualizer.tsx
@@ -3,10 +3,10 @@ import { IconNames } from '@blueprintjs/icons';
 import classNames from 'classnames';
 import * as React from 'react';
 import { configure, GlobalHotKeys } from 'react-hotkeys';
+import DataVisualizer from 'src/features/dataVisualizer/dataVisualizer';
+import { Step } from 'src/features/dataVisualizer/dataVisualizerTypes';
 
-import DataVisualizer from '../../features/dataVisualizer/dataVisualizer';
-import { Step } from '../../features/dataVisualizer/dataVisualizerTypes';
-import { Links } from '../utils/Constants';
+import { Links } from '../../utils/Constants';
 
 type State = {
   steps: Step[];
