+++ src/commons/sideContent/content/githubAssessments/SideContentTestcaseEditor.tsx
@@ -4,9 +4,9 @@ import { Tooltip2 } from '@blueprintjs/popover2';
 import * as React from 'react';
 import AceEditor from 'react-ace';
 
-import { Testcase } from '../../assessment/AssessmentTypes';
-import ControlButton from '../../ControlButton';
-import { showSimpleConfirmDialog } from '../../utils/DialogHelper';
+import { Testcase } from '../../../assessment/AssessmentTypes';
+import ControlButton from '../../../ControlButton';
+import { showSimpleConfirmDialog } from '../../../utils/DialogHelper';
 import SideContentTestcaseCard from '../SideContentTestcaseCard';
 import SideContentEditableTestcaseCard from './SideContentEditableTestcaseCard';
