+++ src/commons/sideContent/content/remoteExecution/DeviceMenuItemButtons.tsx
