+++ src/commons/sideContent/content/remoteExecution/SideContentRemoteExecution.tsx
@@ -24,9 +24,9 @@ import {
 } from 'src/features/remoteExecution/RemoteExecutionEv3Types';
 import { Device, DeviceSession } from 'src/features/remoteExecution/RemoteExecutionTypes';
 
-import { actions } from '../../utils/ActionsHelper';
-import { useTypedSelector } from '../../utils/Hooks';
-import { WorkspaceLocation } from '../../workspace/WorkspaceTypes';
+import { actions } from '../../../utils/ActionsHelper';
+import { useTypedSelector } from '../../../utils/Hooks';
+import { WorkspaceLocation } from '../../../workspace/WorkspaceTypes';
 import DeviceMenuItemButtons from './DeviceMenuItemButtons';
 
 interface SideContentRemoteExecutionProps {
