+++ src/pages/academy/grading/subcomponents/GradingWorkspace.tsx
@@ -6,7 +6,7 @@ import React, { useEffect, useMemo, useState } from 'react';
 import { useDispatch } from 'react-redux';
 import { useNavigate } from 'react-router';
 import { fetchGrading } from 'src/commons/application/actions/SessionActions';
-import SideContentToneMatrix from 'src/commons/sideContent/SideContentToneMatrix';
+import SideContentToneMatrix from 'src/commons/sideContent/content/SideContentToneMatrix';
 import { showSimpleErrorDialog } from 'src/commons/utils/DialogHelper';
 import { useTypedSelector } from 'src/commons/utils/Hooks';
 import {
@@ -51,8 +51,8 @@ import { ControlBarRunButton } from '../../../../commons/controlBar/ControlBarRu
 import { convertEditorTabStateToProps } from '../../../../commons/editor/EditorContainer';
 import { Position } from '../../../../commons/editor/EditorTypes';
 import Markdown from '../../../../commons/Markdown';
+import SideContentAutograder from '../../../../commons/sideContent/content/SideContentAutograder';
 import { SideContentProps } from '../../../../commons/sideContent/SideContent';
-import SideContentAutograder from '../../../../commons/sideContent/SideContentAutograder';
 import { SideContentTab, SideContentType } from '../../../../commons/sideContent/SideContentTypes';
 import Workspace, { WorkspaceProps } from '../../../../commons/workspace/Workspace';
 import { WorkspaceLocation, WorkspaceState } from '../../../../commons/workspace/WorkspaceTypes';
