+++ src/pages/academy/sourcereel/Sourcereel.tsx
@@ -42,8 +42,8 @@ import {
   SourcecastEditorContainerProps
 } from '../../../commons/editor/EditorContainer';
 import { Position } from '../../../commons/editor/EditorTypes';
-import SideContentDataVisualizer from '../../../commons/sideContent/SideContentDataVisualizer';
-import SideContentEnvVisualizer from '../../../commons/sideContent/SideContentEnvVisualizer';
+import SideContentDataVisualizer from '../../../commons/sideContent/content/SideContentDataVisualizer';
+import SideContentEnvVisualizer from '../../../commons/sideContent/content/SideContentEnvVisualizer';
 import { SideContentTab, SideContentType } from '../../../commons/sideContent/SideContentTypes';
 import SourceRecorderControlBar, {
   SourceRecorderControlBarProps
