+++ src/pages/sourcecast/Sourcecast.tsx
@@ -55,8 +55,8 @@ import {
 import MobileWorkspace, {
   MobileWorkspaceProps
 } from '../../commons/mobileWorkspace/MobileWorkspace';
-import SideContentDataVisualizer from '../../commons/sideContent/SideContentDataVisualizer';
-import SideContentEnvVisualizer from '../../commons/sideContent/SideContentEnvVisualizer';
+import SideContentDataVisualizer from '../../commons/sideContent/content/SideContentDataVisualizer';
+import SideContentEnvVisualizer from '../../commons/sideContent/content/SideContentEnvVisualizer';
 import { SideContentTab, SideContentType } from '../../commons/sideContent/SideContentTypes';
 import SourceRecorderControlBar, {
   SourceRecorderControlBarProps
